change navbar to have user dropdown with profile and logout

+++ app/controllers/admin_controller.rb
@@ -5,11 +5,13 @@ class AdminController < ApplicationController
   before_action :authenticate_user!
 
   def clear_db
-    Device.delete_all
-    MosquittoAccount.delete_all
-    MosquittoAcl.delete_all
-    ProvisionRequest.delete_all
+    unless Rails.env.development?
+      Device.delete_all
+      MosquittoAccount.delete_all
+      MosquittoAcl.delete_all
+      ProvisionRequest.delete_all
 
-    redirect_to provision_requests_path
+      redirect_to provision_requests_path
+    end
   end
 end
